// src/@types/AddleadModal.d.ts
export type AddleadlistnameProps={
    name:string,
    remarks:string,
    phone_country_code:string,
    landline_number:string,
    whatsapp_country_code:string,
    alter_country_code:string,
    company_name: string,
    contact_person: string,
    address: string,
    area: string,
    phone: string,
    email: string,
    alternative_no: string,
    whatsapp_no: string,
    customer_category_id: string,
    enquiry_type_id: string,
    requirements_id: string,
    state: string,
    country: string,
    city: string,
    dealer_id:string,
    assignedTo:string,
    receivedDate:string,
    referedBy:string,
    referedPhone:string,
    refer_country_code:string,
    notes:string,
    description:string,
    isNew: string,
    latitude: string,
    longitude: string,
    customerId: string,
    Pincode: string,
    schedule_date: string,
    upload_file: string,
    approximate_amount: string,
}

export type requirementdropdownProps={
    RequirementsId:string,
    RequirementsName:string
}
export type dealerdropdownProps={
    userId:string,
    userName:string,
}